Fix site update count

+++ src/storage/sites_store.py
@@ -21,19 +21,45 @@ class SitesStore:
         self.location = location
         self.fs_client = firestore.Client()
 
-    def update(self, sites: [Site]):
+    def update(self, sites: [Site]) -> int:
+        """
+        Updates Firestore collection with new sites
+
+        Args:
+            sites: list of Site objects
+
+        Returns:
+            No of documents updated in this run
+
+        """
+        total_sites_updated = 0
+        grouped_by_date = itertools.groupby(sorted(sites, key=_grouping_key), key=_grouping_key)
+
         location_col_ref = self.fs_client.collection(self.location)
 
-        grouped_by_date = itertools.groupby(sites, key=lambda s: as_string(s.added_time))
         for date, daily_sites in grouped_by_date:
             sites_as_list = list(daily_sites)
+
+            date_doc_count = 0
             date_doc_ref = location_col_ref.document(date)
+            if date_doc_ref.get().exists:
+                date_doc_count = int(date_doc_ref.get().to_dict()['count'])
 
-            # delete doc if it exists
-            date_doc_ref.delete()
+            if date_doc_count != len(sites_as_list):
+                # this means no of sites on this date has updated
 
-            self.batch_write(date_doc_ref, sites_as_list)
-            date_doc_ref.set({'count': len(sites_as_list)})
+                # ideally, we should check for existence of each doc in the subcollection
+                # but that's too much complexity for our use case. so, just delete/recreate the doc
+                date_doc_ref.delete()
+
+                self.batch_write(date_doc_ref, sites_as_list)
+                date_doc_ref.set({'count': len(sites_as_list)})
+
+                updated_sites = len(sites_as_list) - date_doc_count
+                total_sites_updated += updated_sites
+                print(f'{updated_sites} new sites found for date {date}')
+
+        return total_sites_updated
 
     def batch_write(self, doc, sites: [Site]):
         batch = self.fs_client.batch()
@@ -45,3 +71,7 @@ class SitesStore:
             batch.set(site_doc_ref, site.to_dict())
 
         batch.commit()
+
+
+def _grouping_key(site: Site):
+    return as_string(site.added_time)
